Restructure view and screens packages
