+++ app/src/main/java/nz/eloque/foss_wallet/MainActivity.kt
@@ -27,8 +27,8 @@ import nz.eloque.foss_wallet.persistence.InvalidPassException
 import nz.eloque.foss_wallet.persistence.PassLoader
 import nz.eloque.foss_wallet.shortcut.Shortcut
 import nz.eloque.foss_wallet.ui.WalletApp
+import nz.eloque.foss_wallet.ui.screens.wallet.PassViewModel
 import nz.eloque.foss_wallet.ui.theme.WalletTheme
-import nz.eloque.foss_wallet.ui.view.wallet.PassViewModel
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
