+++ app/src/main/java/nz/eloque/foss_wallet/ui/card/PassRows.kt
@@ -20,7 +20,7 @@ import nz.eloque.foss_wallet.model.field.PassField
 import nz.eloque.foss_wallet.ui.components.AbbreviatingText
 import nz.eloque.foss_wallet.ui.components.DateView
 import nz.eloque.foss_wallet.ui.components.LocationButton
-import nz.eloque.foss_wallet.ui.view.pass.HeaderFieldsView
+import nz.eloque.foss_wallet.ui.screens.pass.HeaderFieldsView
 import java.io.File
 
 @Composable
