+++ app/src/main/java/nz/eloque/foss_wallet/ui/components/GroupCard.kt
@@ -34,7 +34,7 @@ import kotlinx.coroutines.launch
 import nz.eloque.foss_wallet.R
 import nz.eloque.foss_wallet.model.Pass
 import nz.eloque.foss_wallet.ui.card.ShortPassCard
-import nz.eloque.foss_wallet.ui.view.wallet.PassViewModel
+import nz.eloque.foss_wallet.ui.screens.wallet.PassViewModel
 import nz.eloque.foss_wallet.utils.darken
 
 @Composable
