+++ app/src/main/java/nz/eloque/foss_wallet/ui/screens/about/AboutScreen.kt
@@ -1,4 +1,4 @@
-package nz.eloque.foss_wallet.ui.screens
+package nz.eloque.foss_wallet.ui.screens.about
 
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.runtime.Composable
@@ -6,7 +6,6 @@ import androidx.compose.ui.res.stringResource
 import androidx.navigation.NavHostController
 import nz.eloque.foss_wallet.R
 import nz.eloque.foss_wallet.ui.WalletScaffold
-import nz.eloque.foss_wallet.ui.view.about.AboutView
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
