+++ app/src/main/java/nz/eloque/foss_wallet/ui/screens/about/AboutView.kt
@@ -1,4 +1,4 @@
-package nz.eloque.foss_wallet.ui.view.about
+package nz.eloque.foss_wallet.ui.screens.about
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
@@ -9,8 +9,8 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Construction
 import androidx.compose.material.icons.filled.Balance
+import androidx.compose.material.icons.filled.Construction
 import androidx.compose.material.icons.filled.PrivacyTip
 import androidx.compose.material.icons.filled.Source
 import androidx.compose.material.icons.filled.Wallet
