+++ app/src/main/java/nz/eloque/foss_wallet/ui/screens/pass/PassShareButton.kt
@@ -1,4 +1,4 @@
-package nz.eloque.foss_wallet.ui.view.pass
+package nz.eloque.foss_wallet.ui.screens.pass
 
 import android.content.Context
 import android.content.Intent
