+++ app/src/main/java/nz/eloque/foss_wallet/ui/screens/settings/SettingsScreen.kt
@@ -1,4 +1,4 @@
-package nz.eloque.foss_wallet.ui.screens
+package nz.eloque.foss_wallet.ui.screens.settings
 
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
@@ -9,8 +9,6 @@ import androidx.navigation.NavHostController
 import nz.eloque.foss_wallet.R
 import nz.eloque.foss_wallet.ui.Screen
 import nz.eloque.foss_wallet.ui.WalletScaffold
-import nz.eloque.foss_wallet.ui.view.settings.SettingsView
-import nz.eloque.foss_wallet.ui.view.settings.SettingsViewModel
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
