+++ app/src/main/java/nz/eloque/foss_wallet/ui/screens/wallet/WalletScreen.kt
@@ -1,4 +1,4 @@
-package nz.eloque.foss_wallet.ui.screens
+package nz.eloque.foss_wallet.ui.screens.wallet
 
 import android.widget.Toast
 import androidx.activity.compose.rememberLauncherForActivityResult
@@ -36,8 +36,6 @@ import nz.eloque.foss_wallet.model.Pass
 import nz.eloque.foss_wallet.persistence.InvalidPassException
 import nz.eloque.foss_wallet.ui.Screen
 import nz.eloque.foss_wallet.ui.WalletScaffold
-import nz.eloque.foss_wallet.ui.view.wallet.PassViewModel
-import nz.eloque.foss_wallet.ui.view.wallet.WalletView
 import nz.eloque.foss_wallet.utils.isScrollingUp
 
 @OptIn(ExperimentalMaterial3Api::class)
