+++ app/src/main/java/nz/eloque/foss_wallet/ui/screens/wallet/WalletView.kt
@@ -1,4 +1,4 @@
-package nz.eloque.foss_wallet.ui.view.wallet
+package nz.eloque.foss_wallet.ui.screens.wallet
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
